back-end/requirements.txt
numpy==1.18.2
aniso8601==8.0.0
asn1crypto==0.24.0
automium==0.2.6
automium-web==0.1.1
Babel==2.6.0
bcrypt==3.1.6
blinker==1.4
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.3.7
cryptography==3.3.2
cycler==0.10.0
decorator==4.4.2
distro==1.3.0
dnspython==1.16.0
email-validator==1.1.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
httplib2==0.19.0
idna==2.6
importlib-metadata==1.6.1
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.3
jsonpickle==1.4.1
keyring==18.0.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.4.1
legacy==0.1.6
lxml==4.6.3
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==3.2.1
MyApplication==0.1.0
netifaces==0.10.4
oauthlib==2.1.0
obspy==1.2.1
olefile==0.46
paramiko==2.6.0
passlib==1.7.1
pbr==5.4.5
pexpect==4.6.0
Pillow==8.1.1
pony==0.7.13
protobuf==3.6.1
psutil==5.6.6
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.2
pycparser==2.20
pycrypto==2.6.1
pyinotify==0.9.6
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.7.3
python-debian==0.1.36
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.2
pyxdg==0.26
PyYAML==5.4
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
scipy==1.4.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
speaklater==1.3
SQLAlchemy==1.2.18
sqlparse==0.2.4
sshtunnel==0.1.4
testresources==2.0.1
urllib3==1.24.2
wadllib==1.3.3
Werkzeug==1.0.1
wget==3.2
WTForms==2.2.1
zipp==3.1.0